+++ packages/stream-metadata/src/handleHealthCheckRequest.ts
